MetricMatrix/holisticPrototypeAPI/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.9
django-guardian==2.4.0
django-oauth-toolkit==1.5.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-guardian==0.3.0
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
pycparser==2.21
pytz==2021.3
requests==2.26.0